Conflicts:
	src/main/webapp/WEB-INF/view/tiles/auctioned_seller.jsp

+++ src/main/java/com/example/artsell/controller/UserController.java
@@ -1,6 +1,9 @@
 package com.example.artsell.controller;
 
+import java.io.PrintWriter;
+
 import javax.servlet.http.HttpServletRequest;
+import javax.servlet.http.HttpServletResponse;
 import javax.servlet.http.HttpSession;
 
 import org.springframework.beans.factory.annotation.Autowired;
@@ -91,16 +94,27 @@ public class UserController {
 
 	// 2차수정
 	@RequestMapping("/user/delete")
-	public String deleteAccount(HttpSession session) throws Exception {
+	public String deleteAccount(HttpSession session, HttpServletResponse response, SessionStatus sessionStatus) throws Exception {
 		UserSession userSession = (UserSession) session.getAttribute("userSession");
 		String userId = userSession.getAccount().getUserId();
-		artSell.deleteAccount(userId);
+		
+		if(artSell.isAuctioningQuit(userId)) {
+			response.setContentType("text/html; charset=UTF-8");
+            PrintWriter out = response.getWriter();
+            out.println("<script>alert('경매중이므로 탈퇴할 수 없습니다.'); history.go(-1);</script>");
+            out.flush();
+		}
+		else {
+			artSell.deleteAccount(userId);
 
-		session.removeAttribute("userSession");
-		session.invalidate();
+			session.removeAttribute("userSession");
+			session.invalidate();
+			sessionStatus.setComplete();
 
-		System.out.println("계정이 삭제되었습니다.");
-		return "main";
+			System.out.println("계정이 삭제되었습니다.");
+		}
+		
+		return "redirect:/home";
 	}
 
 	@RequestMapping("/admin/manage")
@@ -125,7 +139,13 @@ public class UserController {
 	}
 	
 	@RequestMapping("/admin/user_delete")
-	public String deleteUser(@RequestParam("userId") String userId) {
+	public String deleteUser(@RequestParam("userId") String userId, HttpServletResponse response) throws Exception {
+		if(artSell.isAuctioningQuit(userId)) {
+			response.setContentType("text/html; charset=UTF-8");
+            PrintWriter out = response.getWriter();
+            out.println("<script>alert('경매중이므로 탈퇴할 수 없습니다.'); history.go(-1);</script>");
+            out.flush();
+		}
 		artSell.deleteAccount(userId);
 		return "redirect:/admin/manage";
 	}
